Starting real parsers.
checkstyleUpload will scan checkstyle xml and upload to MySQL
classScan will scan the directory supplied for all classes, to add Classes
to the ClassUID table
commitUpload will analyze the commit and upload it to the commit Table
coverageUpload will parse coverage and upload to the coverage Table
findbugs ... you get the point on this one.
methodScan is like class Scan, but does scans for methods.
pmdUpload is, as above.
Pull_test is going to be deleted next commit.
testScan will scan the testDirectory for tests specifically.  Probably can
be wrapped under methodScan and do some CLA funk.